# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat API",
    "description": "Multi-LLM chat service API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/user_chat_model_privilege": {
      "get": {
        "operationId": "listUserChatModelPrivileges",
        "summary": "List user model privileges",
        "tags": [
          "Admin model privileges"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "fullName": {
                        "type": "string"
                      },
                      "userEmail": {
                        "type": "string"
                      },
                      "chatModelName": {
                        "type": "string"
                      },
                      "rateLimit": {
                        "type": "integer"
                      }
                    },
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "id",
                      "fullName",
                      "userEmail",
                      "chatModelName",
                      "rateLimit"
                    ]
                  },
                  "type": "array"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "createUserChatModelPrivilege",
        "summary": "Create a user model privilege",
        "tags": [
          "Admin model privileges"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/chatModelPrivilegeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatModelPrivilege"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/user_chat_model_privilege/{id}": {
      "delete": {
        "operationId": "deleteUserChatModelPrivilege",
        "summary": "Delete a user model privilege",
        "tags": [
          "Admin model privileges"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "updateUserChatModelPrivilege",
        "summary": "Update a user model privilege",
        "tags": [
          "Admin model privileges"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/chatModelPrivilegeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatModelPrivilege"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/chat_prompts": {
      "post": {
        "operationId": "createChatPrompt",
        "summary": "Create or get a chat prompt",
        "description": "Creates a prompt with status 201, or returns the existing session system prompt with status 200.",
        "tags": [
          "prompts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/chatPromptRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/promptHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/uuid/chat_prompts/{uuid}": {
      "delete": {
        "operationId": "deleteChatPrompt",
        "summary": "Delete a chat prompt",
        "tags": [
          "prompts"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "updateChatPrompt",
        "summary": "Update a chat prompt",
        "tags": [
          "prompts"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleChatMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/promptHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ChatModelPrivilege": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "id": {
            "type": "integer"
          },
          "fullName": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "chatModelName": {
            "type": "string"
          },
          "rateLimit": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "fullName",
          "userEmail",
          "chatModelName",
          "rateLimit"
        ]
      },
...
      "SimpleChatMessage": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "dateTime": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "inversion": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "loading": {
            "type": "boolean"
          },
          "isPin": {
            "type": "boolean"
          },
          "isPrompt": {
            "type": "boolean"
          },
          "artifacts": {
            "items": {
              "properties": {
                "uuid": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "language": {
                  "type": "string"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "uuid",
                "type",
                "title",
                "content"
              ]
            },
            "type": "array"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "uuid",
          "dateTime",
          "text",
          "inversion",
          "error",
          "loading",
          "isPin",
          "isPrompt"
        ]
      },
...
      "chatModelPrivilegeRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "userEmail": {
            "type": "string"
          },
          "chatModelName": {
            "type": "string"
          },
          "rateLimit": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "userEmail",
          "chatModelName",
          "rateLimit"
        ]
      },
      "chatPromptRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "chatSessionUuid": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tokenCount": {
            "type": "integer"
          },
          "score": {
            "type": "number"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "uuid",
          "chatSessionUuid",
          "role",
          "content",
          "tokenCount",
          "score"
        ]
      },
...
      "promptHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "id": {
            "type": "integer"
          },
          "uuid": {
            "type": "string"
          },
          "chatSessionUuid": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "userId": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer"
          },
          "updatedBy": {
            "type": "integer"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "tokenCount": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "uuid",
          "chatSessionUuid",
          "role",
          "content",
          "score",
          "userId",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy",
          "isDeleted",
          "tokenCount"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Browser JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}